+++ test/flex_sub_themes_test.dart
@@ -483,7 +483,6 @@ void main() {
         FlexSubThemes.popupMenuTheme(),
         equals(
           const PopupMenuThemeData(
-            // elevation: 3,
             shape: RoundedRectangleBorder(
               borderRadius: BorderRadius.all(
                 Radius.circular(4),